Uploaded GameLogic.java
Description:
A class with a game board.

Functionality:
-Add players via the player creation panel
-Move the players a number of tiles in their own panel

Possible Improvements:
-Make a new function that calls the move() method recursively with an argument of 1 tile to make players move 1 tile at a time.
For example, if you wanted to move 8 tiles, you would call the move() function 8 times with an argument of 1 (moving 1 tile at a time).
This way, it looks like the player is moving 1 tile at a time, like in Snakes and Ladders. But the current move method is useful for "teleporting" too.

Do your best to break my move logic.