20191212 Update
Patch Notes 20191212
Added a statement oos.reset before writing the updated board to the Server in the BOARDREQUEST case.
This is because when you send the same object over the ObjectOutputStream, it actually sends a backreference to the old object. By calling reset, you empty the stream's cache so it's forced to send the object it recieves.

Added a new class BoardTile which extends JPanel. It has portal functionality.

The move() function in GameLogic now has an additional parameter: boolean ignorePortals.
This is handy for when you're moving one by one and you don't want to trip every portal you pass, just the last tile.

Moved the moving logic to a new method: getEndPosition()
The move method now calls this method to determine the arguments passed to addToBoard()
I was originally planning on using getEndPosition() to evaluate portal state, but I forgot that the moveOneByOne method actually calls move(1), meaning that it doesn't really matter.
I decided to keep this method for code readability.

Idea: For turn control, the client can send a TurnFinished ControlToken so the server knows when to enable the next player's button.