Changelog 2019-12-13
### Added 
- TurnHandler thread on the Server. Controls turns according to the queue implementation. Controlled via a new type of ControlToken, TURNFINISHEDCODE.

### Fixed
- When your roll quantity will overshoot the last tile, instead of proceeding with the move() call quantity# of times, display a JOptionPane explaining
that you failed to land exactly on the last tile.

### Fixed
- JOptionPane messages, specifically for portals, now only display for the current player (the moving player). This is done by adding an additional parameter,
invokingPlayer, to the move methods. So when move() is called in the client, it will compare invokingPlayer and the moving player to decide whether or not to
display the message or not.

### Changed
- Migrated the console command functionality to GUI. All incoming messages and commands are displayed in the GUI now.

### Added
- Added new commands, getboard, and showboard. getboard will request an updated board from the first client in the Vector. showboard will display that board in
a new window.

### Fixed
- Pawns will no longer vary across clients when added to the board. Instead of deciding color in the Player constructor, a random int from 0-7 is generated on
the Server upon client connection, which is sent to the clients for construction of a player. Because the int is the same for all clients, the pawn will always
be the same color.