Committing my 12-13 to 12-18 code updates
Changelog TBA

### Fixed
- When clients disconnnect, they are removed from the queue (IMPORTANT)!!!
- Fix client same name disconnection causing errors

Need to allow server and clients to choose board size? Nah
Changelog 2019-12-18
### Added
- Auto-scroll functionality to the game messages JTextArea.

Changelog 2019-12-17
### Changed
- Instead of serializing an entire GameLogic, serialize the instructions to create an updated board. This way, avoid issues with serialVersionUID due to
Java components not being serialiazable (at least not without issues)

### Changed
- The next player in the queue can not roll until the moving player has finished moving.

### Changed
- Add the wait() + notify() mechanism to ControlToken.TURNFINISHEDCODE. The Server will wait for the current player to finish moving before enabling the next player's die roll button.

### Deprecated
- RollRequest was merged into ControlToken. There was no need for it to be a separate class.

### Added
- A main menu with PLAY, HELP, and EXIT.

### Fixed
- When a client overshoots, the next player's turn is now enabled. This was because I forgot to add turnFinished = true in the overshoot case.

### Added
- A scroll bar to the server command line.

### Changed
- The win message displayed to everyone when somebody lands on the last tile now closes the game when "OK" is clicked. This is the Server telling the Clients
that the game is done.

Changelog 2019-12-16
### Changed
- Instead of using a timer to wait for a new board to be sent to the server (which would result in unnecessary wait time),
the client thread will wait() until it recieves a new board. Then it will notify. This way, it is guaranteed for an updated board to be recieved
before sending to the client. 
Side note: I tried using wait + notify previously but it got stuck in the wait loop because I was asking the first client to wait until
it received a new board from itself, which was not possible since it was waiting, meaning it never reached the case statements where the notify statement is.
If you remove the if(clientThreads.get(0) != this) clause, the first client is stuck waiting and never gets to send updated boards.

### Changed
- The JOptionPane now displays before the player moves to the next/previous portal. This means players can see the player's starting portal instead of magically
teleporting to the next/previous portal. This might not fix the issue for clients who don't recieve JOptionPanes (i.e. all other players who aren't currently moving)

### Fixed
- The pawn images are now named 0, 1, 2 ,3 ... 7. This fixes the issue where the client would try to grab the "media/0.png" pawn image and wouldn't
be able to find the file. This is because Random.nextInt(bound) produces a number from 0 inclusive to bound exclusive.

### Removed
- The wait before the first client sends an updated board in the case ControlToken.BOARDREQUEST block. There is no longer a need to wait because the issue wasn't that
the client wasn't having enough time to add the new player, but that the client was writing the same old cached board.

Changelog 2019-12-14
### Added
- Automatic pawn and portal resizing depending on the board size. I had to wrap the portal image in an ImageIcon because you can't serialize a BufferedImage.

### Fixed
- Players now reliably move to the previous orange portal/ next blue portal because of a new function, nextPortal().
- Movement between portals is no longer based on ArrayList positions. Instead, a method in GameLogic will search through the board until it finds a matching portal.

Changelog 2019-12-13
### Added 
- TurnHandler thread on the Server. Controls turns according to the queue implementation. Controlled via a new type of ControlToken, TURNFINISHEDCODE.

### Fixed
- When your roll quantity will overshoot the last tile, instead of proceeding with the move() call quantity# of times, display a JOptionPane explaining
that you failed to land exactly on the last tile.

### Fixed
- JOptionPane messages, specifically for portals, now only display for the current player (the moving player). This is done by adding an additional parameter,
invokingPlayer, to the move methods. So when move() is called in the client, it will compare invokingPlayer and the moving player to decide whether or not to
display the message or not.

### Changed
- Migrated the console command functionality to GUI. All incoming messages and commands are displayed in the GUI now.

### Added
- Added new commands, getboard, and showboard. getboard will request an updated board from the first client in the Vector. showboard will display that board in
a new window.

### Fixed
- Pawns will no longer vary across clients when added to the board. Instead of deciding color in the Player constructor, a random int from 0-7 is generated on
the Server upon client connection, which is sent to the clients for construction of a player. Because the int is the same for all clients, the pawn will always
be the same color.